Styles gallery view and largeImage in individual view
    - Review itself hasn't been styled; just the images
